frontend: minor refactor to article injector
